+++ packages/pty-tunnel-server/scripts/build.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-set -e
-
-sed "s@{{VERCEL_URL}}@$VERCEL_URL@" templates/install > public/install